fix: remove unnecessary source generation
- Moved OptifyAttribute to it's own file
- Moved UseOptify<T> to a separate extension file.

Issue: #12
