+++ Optify/OptifyRegistrationSource.cs
@@ -3,39 +3,9 @@ using System.Text;
 
 namespace Optify;
 
-public static class HostBuilderExtensionsSource
+public static class OptifyRegistrationSource
 {
     public const string Filename = "OptifyRegistration.g.cs";
-
-    private const string UseOptifyGenericExtensionSource =
-        """
-            /// <summary>
-            /// Register a single type options type using the conventional name or one specified on the
-            /// <see cref="OptifyAttribute"/> attribute. Using this method will require reflection during runtime.
-            /// </summary>
-            /// <param name="hostBuilder">The <see cref="IHostBuilder"/> instance to extend.</param>
-            /// <typeparam name="T">The type to map configured options to.</typeparam>
-            /// <returns>The extended <see cref="IHostBuilder"/> instance.</returns>
-            public static IHostBuilder UseOptify<T>(this IHostBuilder hostBuilder) where T : class, new()
-            {
-                hostBuilder.ConfigureServices((ctx, services) =>
-                {
-                    var type = typeof(T);
-                    var attribute = type
-                        .GetCustomAttributes(typeof(OptifyAttribute), false)
-                        .FirstOrDefault(attr => attr is OptifyAttribute);
-            
-                    var sectionName = attribute is OptifyAttribute { SectionName.Length: > 0 } attr
-                        ? attr.SectionName
-                        : type.Name;
-            
-                    services
-                        .AddOptions<T>()
-                        .Bind(ctx.Configuration.GetSection(sectionName));
-                });
-                return hostBuilder;
-            }
-        """;
 
     public static string GenerateSource(ImmutableArray<OptionsTypeToRegister> optionsToRegister)
     {
@@ -69,10 +39,8 @@ public static class HostBuilderExtensionsSource
                 
                     return hostBuilder;
                 }
+            }
             """);
-        sb.AppendLine();
-        sb.AppendLine(UseOptifyGenericExtensionSource);
-        sb.AppendLine("}");
 
         return sb.ToString();
     }
